updating sway configuration override subheading
In the current version, it was stated "sway is a replacement for the popular i3 windows manager in Windows" . In which according to my knowledge the word "Windows" shall be replaced by "Linux" since window managers are much more common and used in Linux (or Unix based systems to be precise). I hope this will help.